refactor(perf): split /workshops + /labs into server shell + motion island
39th + 40th in the pattern series.
